(cxd4) Start converting signed longs to int32_t - Conker's Bad Fur Day no ongre crashes at startup with cxd4

+++ mupen64plus-rsp-cxd4/su.h
@@ -19,7 +19,7 @@
  * abandon their designated purposes on the master CPU host (the VR4300),
  * hence most of the MIPS names "k0, k1, t0, t1, v0, v1 ..." no longer apply.
  */
-static int SR[32];
+static int32_t SR[32];
 
 NOINLINE static void res_S(void)
 {
@@ -434,7 +434,7 @@ static union {
 
 static void LB(void) /* 100000 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const signed int offset = (signed short)(inst.I.imm);
 
     addr = BES(SR[inst.I.rs] + offset) & 0x00000FFF;
@@ -444,7 +444,7 @@ static void LB(void) /* 100000 sssss ttttt iiiiiiiiiiiiiiii */
 }
 static void LH(void) /* 100001 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -464,10 +464,10 @@ static void LH(void) /* 100001 sssss ttttt iiiiiiiiiiiiiiii */
     SR[0] = 0x00000000;
     return;
 }
-extern void ULW(int rd, unsigned long addr);
+extern void ULW(int rd, uint32_t addr);
 static void LW(void) /* 100011 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -477,13 +477,13 @@ static void LW(void) /* 100011 sssss ttttt iiiiiiiiiiiiiiii */
         ULW(rt, addr); /* Address Error exception:  RSP bypass MIPS pseudo-op */
         return;
     }
-    SR[rt] = *(long *)(RSP.DMEM + addr);
+    SR[rt] = *(int32_t*)(RSP.DMEM + addr);
     SR[0] = 0x00000000;
     return;
 }
 static void LBU(void) /* 100100 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const signed int offset = (signed short)(inst.I.imm);
 
     addr = BES(SR[inst.I.rs] + offset) & 0x00000FFF;
@@ -493,7 +493,7 @@ static void LBU(void) /* 100100 sssss ttttt iiiiiiiiiiiiiiii */
 }
 static void LHU(void) /* 100101 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -511,7 +511,7 @@ static void LHU(void) /* 100101 sssss ttttt iiiiiiiiiiiiiiii */
 }
 static void SB(void) /* 101000 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const signed int offset = (signed short)(inst.I.imm);
 
     addr = BES(SR[inst.I.rs] + offset) & 0x00000FFF;
@@ -520,7 +520,7 @@ static void SB(void) /* 101000 sssss ttttt iiiiiiiiiiiiiiii */
 }
 static void SH(void) /* 101001 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -535,10 +535,10 @@ static void SH(void) /* 101001 sssss ttttt iiiiiiiiiiiiiiii */
     *(short *)(RSP.DMEM + addr - HES(0x000)*(addr%4 - 1)) = (short)(SR[rt]);
     return;
 }
-extern void USW(int rs, unsigned long addr);
+extern void USW(int rs, uint32_t addr);
 static void SW(void) /* 101011 sssss ttttt iiiiiiiiiiiiiiii */
 {
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -548,7 +548,7 @@ static void SW(void) /* 101011 sssss ttttt iiiiiiiiiiiiiiii */
         USW(rt, addr); /* Address Error exception:  RSP bypass MIPS pseudo-op */
         return;
     }
-    *(long *)(RSP.DMEM + addr) = SR[rt];
+    *(int32_t*)(RSP.DMEM + addr) = SR[rt];
     return;
 }
 
@@ -944,7 +944,7 @@ static void C2(void)
 /*** Scalar, Coprocessor Operations (vector unit, scalar cache transfers) ***/
 void LS_Group_I(int direction, int length)
 { /* Group I vector loads and stores, as defined in SGI's patent. */
-    register signed long addr;
+    register int32_t addr;
     register int i;
     register int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = SE(inst.SW, 6);
@@ -969,7 +969,7 @@ static void LSV(void)
     LS_Group_I(0, sizeof(short) > 2 ? 2 : sizeof(short));
     return;
 #else
-    register signed long addr;
+    register int32_t addr;
     const int vt   = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = SE(inst.SW, 6);
@@ -992,11 +992,11 @@ static void LSV(void)
 static void LLV(void)
 {
 #if (0)
-    LS_Group_I(0, sizeof(long) > 4 ? 4 : sizeof(long));
+    LS_Group_I(0, sizeof(int32_t) > 4 ? 4 : sizeof(int32_t));
     return;
 #else
     int correction;
-    register signed long addr;
+    register int32_t addr;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = SE(inst.SW, 6);
@@ -1025,7 +1025,7 @@ static void LDV(void)
     LS_Group_I(0, 8);
     return;
 #else
-    register signed long addr;
+    register int32_t addr;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = SE(inst.SW, 6);
@@ -1122,7 +1122,7 @@ static void SSV(void)
     LS_Group_I(1, sizeof(short) > 2 ? 2 : sizeof(short));
     return;
 #else
-    register signed long addr;
+    register int32_t addr;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const int vt = inst.R.rt;
     const signed int offset = SE(inst.SW, 6);
@@ -1145,11 +1145,11 @@ static void SSV(void)
 static void SLV(void)
 {
 #if (0)
-    LS_Group_I(1, sizeof(long) > 4 ? 4 : sizeof(long));
+    LS_Group_I(1, sizeof(int32_t) > 4 ? 4 : sizeof(int32_t));
     return;
 #else
     int correction;
-    register signed long addr;
+    register int32_t addr;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = SE(inst.SW, 6);
@@ -1178,7 +1178,7 @@ static void SDV(void)
     LS_Group_I(1, 8);
     return;
 #else
-    register signed long addr;
+    register int32_t addr;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = SE(inst.SW, 6);
@@ -1267,7 +1267,7 @@ static void SDV(void)
  */
 static void LPV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     register int b;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
@@ -1381,7 +1381,7 @@ static void LPV(void)
 }
 static void LUV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     register int b;
     const int vt = inst.R.rt;
     int e = (inst.W & 0x000007FF) >> (6 + 1); /* fixme ? */ /* inst.R.sa >> 1 */
@@ -1504,7 +1504,7 @@ static void LUV(void)
 static void SPV(void)
 {
     register int b;
-    register signed long addr;
+    register int32_t addr;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const int vt = inst.R.rt;
     const signed int offset = -(inst.SW & 0x00000040) | inst.R.func;
@@ -1618,7 +1618,7 @@ static void SPV(void)
 static void SUV(void)
 {
     register int b;
-    register signed long addr;
+    register int32_t addr;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const int vt = inst.R.rt;
     const signed int offset = -(inst.SW & 0x00000040) | inst.R.func;
@@ -1667,7 +1667,7 @@ static void SUV(void)
  */
 static void LHV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = -(inst.SW & 0x00000040) | inst.R.func;
@@ -1715,7 +1715,7 @@ static void LFV(void)
 }
 static void SHV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const int vt = inst.R.rt;
     const signed int offset = -(inst.SW & 0x00000040) | inst.R.func;
@@ -1744,7 +1744,7 @@ static void SHV(void)
 }
 static void SFV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const int vt = inst.R.rt;
     const signed int offset = -(inst.SW & 0x00000040) | inst.R.func;
@@ -1778,7 +1778,7 @@ static void SFV(void)
  */
 static void LQV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     register int b;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* Boss Game Studios trap */
@@ -1855,7 +1855,7 @@ static void LQV(void)
 }
 static void LRV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     register int b;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* Boss Game Studios trap */
@@ -1924,7 +1924,7 @@ static void LRV(void)
 }
 static void SQV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     register int b;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
@@ -1984,7 +1984,7 @@ static void SQV(void)
 }
 static void SRV(void)
 {
-    register signed long addr;
+    register int32_t addr;
     register int b;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const int vt = inst.R.rt;
@@ -2059,7 +2059,7 @@ static void SRV(void)
 static void LTV(void)
 {
     register int i;
-    register signed long addr;
+    register int32_t addr;
     const int vt = inst.R.rt;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const signed int offset = -(inst.SW & 0x00000040) | inst.R.func;
@@ -2105,7 +2105,7 @@ static void SWV(void)
 static void STV(void)
 {
     register int i;
-    register signed long addr;
+    register int32_t addr;
     const int e = (inst.W & 0x000007FF) >> (6 + 1); /* inst.R.sa >> 1 */
     const int vt = inst.R.rt;
     const signed int offset = -(inst.SW & 0x00000040) | inst.R.func;
@@ -2178,7 +2178,7 @@ static void BNEZ(void)
     set_PC(*RSP.SP_PC_REG + 4*offset + SLOT_OFF);
     return;
 }
-void ULW(int rd, unsigned long addr)
+void ULW(int rd, uint32_t addr)
 { /* "Unaligned Load Word" */
     if (addr & 0x00000001)
     {
@@ -2200,7 +2200,7 @@ void ULW(int rd, unsigned long addr)
  /* SR[0] = 0x00000000; */
     return;
 }
-void USW(int rs, unsigned long addr)
+void USW(int rs, uint32_t addr)
 { /* "Unaligned Store Word" */
     SR_temp.W = SR[rs];
     if (addr & 0x00000001)
@@ -2233,7 +2233,7 @@ static void LXI(void)
 } /* should translate to `ADDI[U] rt, $zero, imm` */
 static void LBA(void)
 { /* "Load Byte from Absolute Address" (unofficial, created by me) */
-    register signed long addr;
+    register int32_t addr;
     const signed int offset = (signed short)(inst.I.imm);
 
     addr = BES(0x00000000 + offset) & 0x00000FFF;
@@ -2243,7 +2243,7 @@ static void LBA(void)
 }
 static void LHA(void)
 { /* "Load Halfword from Absolute Address" (unofficial, created by me) */
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -2262,7 +2262,7 @@ static void LHA(void)
 }
 static void LWA(void)
 { /* "Load Word from Absolute Address" (unofficial, created by me) */
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -2272,13 +2272,13 @@ static void LWA(void)
         ULW(rt, addr); /* Address Error exception:  RSP bypass MIPS pseudo-op */
         return;
     }
-    SR[rt] = *(long *)(RSP.DMEM + addr);
+    SR[rt] = *(int32_t*)(RSP.DMEM + addr);
     SR[0] = 0x00000000;
     return;
 }
 static void LBUA(void)
 { /* "Load Byte Unsigned from Absolute Address" (unofficial, created by me) */
-    register signed long addr;
+    register int32_t addr;
     const signed int offset = (signed short)(inst.I.imm);
 
     addr = BES(0x00000000 + offset) & 0x00000FFF;
@@ -2288,7 +2288,7 @@ static void LBUA(void)
 }
 static void LHUA(void)
 { /* "Load Halfword Unsigned from Absolute Address" (unofficial...) */
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -2306,7 +2306,7 @@ static void LHUA(void)
 }
 static void SBA(void)
 { /* "Store Byte from Absolute Address" (unofficial, created by me) */
-    register signed long addr;
+    register int32_t addr;
     const signed int offset = (signed short)(inst.I.imm);
 
     addr = BES(0x00000000 + offset) & 0x00000FFF;
@@ -2315,7 +2315,7 @@ static void SBA(void)
 }
 static void SHA(void)
 { /* "Store Halfword from Absolute Address" (unofficial, created by me) */
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -2327,12 +2327,12 @@ static void SHA(void)
         RSP.DMEM[addr + BES(0x000)] = SR_B(rt, 3);
         return;
     }
-    *(short *)(RSP.DMEM + addr - HES(0x000)*(addr%4 - 1)) = (short)(SR[rt]);
+    *(int16_t*)(RSP.DMEM + addr - HES(0x000)*(addr%4 - 1)) = (short)(SR[rt]);
     return;
 }
 static void SWA(void)
 { /* "Store Word from Absolute Address" (unofficial, created by me) */
-    register signed long addr;
+    register int32_t addr;
     const int rt = inst.I.rt;
     const signed int offset = (signed short)(inst.I.imm);
 
@@ -2342,7 +2342,7 @@ static void SWA(void)
         USW(rt, addr); /* Address Error exception:  RSP bypass MIPS pseudo-op */
         return;
     }
-    *(long *)(RSP.DMEM + addr) = SR[rt];
+    *(int32_t*)(RSP.DMEM + addr) = SR[rt];
     return;
 }
